backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic>=2.9.0
supabase>=2.0.0
google-genai>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.2.0
requests>=2.31.0
Faker>=20.0.0
google-auth>=2.27.0
